chore: merge code from `monaco.ts` to `EditorUI.ts`
