+++ spx-gui/src/components/editor/code-editor/tools/common.ts
