+++ spx-gui/src/components/editor/code-editor/tools/gop.ts
