+++ spx-gui/src/components/editor/code-editor/tools/index.ts
