+++ spx-gui/src/components/editor/code-editor/ui/code-text-editor/index.ts
@@ -1,2 +1,1 @@
 export { default as CodeTextEditor } from './CodeTextEditor.vue'
-export * from './tools'
